+++ rust/zkgroup/src/api/auth/auth_credential_with_pni/zkc.rs
@@ -57,7 +57,23 @@ impl AuthCredentialWithPniZkcResponse {
     ) -> Self {
         Self::issue_credential_for_key(
             aci,
-            pni,
+            UidStruct::from_service_id(pni.into()),
+            redemption_time,
+            &params.generic_credential_key_pair,
+            randomness,
+        )
+    }
+
+    pub fn issue_credential_without_pni(
+        aci: Aci,
+        pni_salt: &[u8],
+        redemption_time: Timestamp,
+        params: &ServerSecretParams,
+        randomness: RandomnessBytes,
+    ) -> Self {
+        Self::issue_credential_for_key(
+            aci,
+            UidStruct::for_false_pni_derived_from_aci(aci, pni_salt),
             redemption_time,
             &params.generic_credential_key_pair,
             randomness,
@@ -73,22 +89,37 @@ impl AuthCredentialWithPniZkcResponse {
     ) -> Result<AuthCredentialWithPniZkc, ZkGroupVerificationFailure> {
         self.receive_for_key(
             aci,
-            pni,
+            UidStruct::from_service_id(pni.into()),
             redemption_time,
             &public_params.generic_credential_public_key,
         )
     }
 
-    pub(crate) fn issue_credential_for_key(
+    pub fn receive_without_pni(
+        self,
         aci: Aci,
-        pni: Pni,
+        pni_salt: &[u8],
+        redemption_time: Timestamp,
+        public_params: &ServerPublicParams,
+    ) -> Result<AuthCredentialWithPniZkc, ZkGroupVerificationFailure> {
+        self.receive_for_key(
+            aci,
+            UidStruct::for_false_pni_derived_from_aci(aci, pni_salt),
+            redemption_time,
+            &public_params.generic_credential_public_key,
+        )
+    }
+
+    fn issue_credential_for_key(
+        aci: Aci,
+        pni_points: UidStruct,
         redemption_time: Timestamp,
         credential_key: &CredentialKeyPair,
         randomness: RandomnessBytes,
     ) -> Self {
         let proof = zkcredential::issuance::IssuanceProofBuilder::new(CREDENTIAL_LABEL)
             .add_attribute(&UidStruct::from_service_id(aci.into()))
-            .add_attribute(&UidStruct::from_service_id(pni.into()))
+            .add_attribute(&pni_points)
             .add_public_attribute(&redemption_time)
             .issue(credential_key, randomness);
 
@@ -98,10 +129,10 @@ impl AuthCredentialWithPniZkcResponse {
         }
     }
 
-    pub(crate) fn receive_for_key(
+    fn receive_for_key(
         self,
         aci: Aci,
-        pni: Pni,
+        pni_points: UidStruct,
         redemption_time: Timestamp,
         public_key: &CredentialPublicKey,
     ) -> Result<AuthCredentialWithPniZkc, ZkGroupVerificationFailure> {
@@ -110,11 +141,10 @@ impl AuthCredentialWithPniZkcResponse {
         }
 
         let aci = UidStruct::from_service_id(aci.into());
-        let pni = UidStruct::from_service_id(pni.into());
 
         let raw_credential = zkcredential::issuance::IssuanceProofBuilder::new(CREDENTIAL_LABEL)
             .add_attribute(&aci)
-            .add_attribute(&pni)
+            .add_attribute(&pni_points)
             .add_public_attribute(&redemption_time)
             .verify(public_key, self.proof)?;
 
@@ -122,7 +152,7 @@ impl AuthCredentialWithPniZkcResponse {
             credential: raw_credential,
             version: VersionByte,
             aci,
-            pni,
+            pni: pni_points,
             redemption_time,
         })
     }
@@ -243,15 +273,33 @@ mod test {
 
         let response = AuthCredentialWithPniZkcResponse::issue_credential_for_key(
             ACI,
-            PNI,
+            UidStruct::from_service_id(PNI.into()),
             REDEMPTION_TIME,
             &credential_key,
             [3; RANDOMNESS_LEN],
         );
 
         let credential = response
-            .receive_for_key(ACI, PNI, REDEMPTION_TIME, public_key)
+            .clone()
+            .receive_for_key(
+                ACI,
+                UidStruct::from_service_id(PNI.into()),
+                REDEMPTION_TIME,
+                public_key,
+            )
             .expect("is valid");
+
+        assert!(
+            response
+                .receive_for_key(
+                    ACI,
+                    UidStruct::for_false_pni_derived_from_aci(ACI, b"salt"),
+                    REDEMPTION_TIME,
+                    public_key,
+                )
+                .is_err(),
+            "omitted PNI where one was expected"
+        );
 
         let presentation =
             credential.present_for_key(public_key, &group_secret_params, [4; RANDOMNESS_LEN]);
@@ -263,5 +311,70 @@ mod test {
                 REDEMPTION_TIME,
             )
             .expect("can verify")
+    }
+
+    #[test]
+    fn issue_receive_present_without_pni() {
+        const ACI: Aci = Aci::from_uuid_bytes([b'a'; 16]);
+        const REDEMPTION_TIME: Timestamp = Timestamp::from_epoch_seconds(12345 * SECONDS_PER_DAY);
+
+        let credential_key = CredentialKeyPair::generate([1; RANDOMNESS_LEN]);
+        let public_key = credential_key.public_key();
+        let group_secret_params = GroupSecretParams::generate([2; RANDOMNESS_LEN]);
+
+        let response = AuthCredentialWithPniZkcResponse::issue_credential_for_key(
+            ACI,
+            UidStruct::for_false_pni_derived_from_aci(ACI, b"salt"),
+            REDEMPTION_TIME,
+            &credential_key,
+            [3; RANDOMNESS_LEN],
+        );
+
+        let credential = response
+            .clone()
+            .receive_for_key(
+                ACI,
+                UidStruct::for_false_pni_derived_from_aci(ACI, b"salt"),
+                REDEMPTION_TIME,
+                public_key,
+            )
+            .expect("is valid");
+
+        assert!(
+            response
+                .clone()
+                .receive_for_key(
+                    ACI,
+                    UidStruct::for_false_pni_derived_from_aci(ACI, b"pepper"),
+                    REDEMPTION_TIME,
+                    public_key,
+                )
+                .is_err(),
+            "wrong salt"
+        );
+        assert!(
+            response
+                .receive_for_key(
+                    ACI,
+                    UidStruct::from_service_id(Pni::from_uuid_bytes([b'p'; 16]).into()),
+                    REDEMPTION_TIME,
+                    public_key,
+                )
+                .is_err(),
+            "provided a PNI where one was not expected"
+        );
+
+        let presentation =
+            credential.present_for_key(public_key, &group_secret_params, [4; RANDOMNESS_LEN]);
+
+        presentation
+            .verify_for_key(
+                &credential_key,
+                &group_secret_params.get_public_params(),
+                REDEMPTION_TIME,
+            )
+            .expect("can verify");
+
+        assert!(presentation.aci_ciphertext() != presentation.pni_ciphertext());
     }
 }