+++ swift/Tests/LibSignalClientTests/ZKGroupTests.swift
@@ -184,6 +184,70 @@ class ZKGroupTests {
     }
 
     @Test
+    func authZkcWithoutPniIntegration() throws {
+        // Both ACI and salt will be assigned to an account at registration.
+        let aci = Aci(fromUUID: TEST_ARRAY_16)
+        let salt = Data([1, 2, 3, 4, 5])
+        let redemptionTime: UInt64 = 123_456 * SECONDS_PER_DAY
+
+        // Generate keys (client's are per-group, server's are not)
+        // ---
+
+        // SERVER
+        let serverSecretParams = try ServerSecretParams.generate(randomness: self.TEST_ARRAY_32)
+        let serverPublicParams = try serverSecretParams.getPublicParams()
+        let serverZkAuth = ServerZkAuthOperations(serverSecretParams: serverSecretParams)
+
+        // CLIENT
+        let masterKey = try GroupMasterKey(contents: TEST_ARRAY_32_1)
+        let groupSecretParams = try GroupSecretParams.deriveFromMasterKey(groupMasterKey: masterKey)
+        let groupPublicParams = try groupSecretParams.getPublicParams()
+
+        // SERVER
+        // Issue credential
+        let authCredentialResponse = try serverZkAuth.issueAuthCredentialZkcWithoutPni(
+            randomness: self.TEST_ARRAY_32_2,
+            aci: aci,
+            salt: salt,
+            redemptionTime: redemptionTime
+        )
+
+        // CLIENT
+        // Receive credential
+        let clientZkAuthCipher = ClientZkAuthOperations(serverPublicParams: serverPublicParams)
+        let clientZkGroupCipher = ClientZkGroupCipher(groupSecretParams: groupSecretParams)
+        let authCredential = try clientZkAuthCipher.receiveAuthCredentialWithoutPni(
+            aci: aci,
+            salt: salt,
+            redemptionTime: redemptionTime,
+            authCredentialResponse: authCredentialResponse
+        )
+
+        // Create and decrypt user entry
+        let aciCiphertext = try clientZkGroupCipher.encrypt(aci)
+        let aciPlaintext = try clientZkGroupCipher.decrypt(aciCiphertext)
+        #expect(aci == aciPlaintext)
+
+        // Create presentation
+        let presentation = try clientZkAuthCipher.createAuthCredentialPresentation(
+            randomness: self.TEST_ARRAY_32_5,
+            groupSecretParams: groupSecretParams,
+            authCredential: authCredential
+        )
+
+        // Verify presentation
+        let uuidCiphertextRecv = try presentation.getUuidCiphertext()
+        #expect(aciCiphertext.serialize() == uuidCiphertextRecv.serialize())
+        #expect(try presentation.getPniCiphertext().serialize() != aciCiphertext.serialize())
+        #expect(try presentation.getRedemptionTime() == Date(timeIntervalSince1970: TimeInterval(redemptionTime)))
+        try serverZkAuth.verifyAuthCredentialPresentation(
+            groupPublicParams: groupPublicParams,
+            authCredentialPresentation: presentation,
+            now: Date(timeIntervalSince1970: TimeInterval(redemptionTime))
+        )
+    }
+
+    @Test
     func expiringProfileKeyIntegration() throws {
         let userId = Aci(fromUUID: TEST_ARRAY_16)
         // Generate keys (client's are per-group, server's are not)